Abstracts some parts of the cache busting
